光盘库节点详细概况: 完成单个光盘库节点通信后台收发报文处理，测试通过

+++ src/main/java/com/shdd/cfs/dto/device/tape/TapeNodeDetail.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class TapeNodeDetail {
     private Integer id;
     private String name;
-    private Integer used;
-    private Integer capacity;
+    private Double used;
+    private Double capacity;
     private Integer status;
 }
